refactor(user): move usecases and update port with existence check (#3)
