+++ src/main/java/com/lernix/infrastructure/web/controller/UserController.java
@@ -4,7 +4,7 @@ import com.lernix.domain.model.User;
 import com.lernix.infrastructure.web.dto.request.CreateUserRequest;
 import com.lernix.infrastructure.web.dto.response.UserResponse;
 import com.lernix.infrastructure.web.mapper.UserMapper;
-import com.lernix.application.usecase.CreateUserUseCase;
+import com.lernix.application.usecase.user.CreateUserUseCase;
 import com.lernix.shared.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
